third_party/WebKit/LayoutTests/fast/frames/sandboxed-iframe-parsing-space-characters-expected.txt

CONSOLE MESSAGE: line 1: PASS: Form feed is a delimiter.
CONSOLE ERROR: line 27: Error while parsing the 'sandbox' attribute: 'allow-scriptsallow-forms' is an invalid sandbox flag.
CONSOLE ERROR: Blocked script execution in 'data:text/html,<script>console.log('FAIL: Vertical tab is not a delimiter.');</script>' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: line 1: PASS: Newline is a delimiter.
CONSOLE MESSAGE: line 1: PASS: Return is a delimiter.
CONSOLE ERROR: line 27: Error while parsing the 'sandbox' attribute: 'allow-scriptsxallow-forms' is an invalid sandbox flag.
CONSOLE ERROR: Blocked script execution in 'data:text/html,<script>console.log('FAIL: x is not a delimiter.');</script>' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: line 1: PASS: Tab is a delimiter.
CONSOLE MESSAGE: line 1: PASS: Space is a delimiter character.
This tests whether we correct parse various space characters in the sandbox attribute.
 
